+++ lib/karafka/web/ui/helpers/application_helper.rb
@@ -132,13 +132,14 @@ module Karafka
           # @return [String] span tag with label and title with change time if present
           def poll_state_with_change_time_label(state, state_ch)
             year_in_seconds = 131_556_926
+            state_ch_in_seconds = state_ch / 1_000.0
 
             # If state is active, there is no date of change
             if state == 'active'
               %(
                 <span class="badge #{kafka_state_bg(state)} mt-1 mb-1">#{state}</span>
               )
-            elsif Time.at(state_ch) > Time.now + year_in_seconds
+            elsif state_ch_in_seconds > year_in_seconds
               %(
                 <span
                   class="badge #{kafka_state_bg(state)} mt-1 mb-1"
@@ -151,7 +152,7 @@ module Karafka
               %(
                 <span
                   class="badge #{kafka_state_bg(state)} time-title mt-1 mb-1"
-                  title="#{Time.at(state_ch)}"
+                  title="#{Time.now + state_ch_in_seconds}"
                 >
                   #{state}
                 </span>