+++ spec/lib/karafka/web/tracking/consumers/contracts/subscription_group_spec.rb
@@ -28,7 +28,8 @@ RSpec.describe_current do
               ls_offset_d: 0,
               ls_offset_fd: 0,
               fetch_state: 'active',
-              poll_state: 'active'
+              poll_state: 'active',
+              poll_state_ch: 0
             }
           }
         }