+++ spec/lib/karafka/web/ui/pro/controllers/health_spec.rb
@@ -180,4 +180,81 @@ RSpec.describe_current do
       end
     end
   end
+
+  describe '#changes' do
+    context 'when no report data' do
+      before do
+        topics_config.consumers.reports = reports_topic
+
+        get 'health/changes'
+      end
+
+      it do
+        expect(response).to be_ok
+        expect(body).to include(breadcrumbs)
+        expect(body).not_to include(pagination)
+        expect(body).not_to include(support_message)
+        expect(body).to include('No health data is available')
+        expect(body).not_to include('bg-warning')
+        expect(body).not_to include('bg-danger')
+      end
+    end
+
+    context 'when data is present' do
+      before { get 'health/changes' }
+
+      it do
+        expect(response).to be_ok
+        expect(body).to include(breadcrumbs)
+        expect(body).not_to include(pagination)
+        expect(body).not_to include(support_message)
+        expect(body).to include('Pause state change')
+        expect(body).to include('N/A')
+        expect(body).to include('2690818656.575513')
+      end
+    end
+
+    context 'when data is present but reported in a transactional fashion' do
+      before do
+        topics_config.consumers.reports = reports_topic
+        produce(reports_topic, Fixtures.file('consumer_report.json'), type: :transactional)
+
+        get 'health/changes'
+      end
+
+      it do
+        expect(response).to be_ok
+        expect(body).to include(breadcrumbs)
+        expect(body).not_to include(pagination)
+        expect(body).not_to include(support_message)
+        expect(body).to include('Pause state change')
+        expect(body).to include('Changes')
+      end
+    end
+
+    context 'when one of partitions is paused forever' do
+      before do
+        topics_config.consumers.reports = reports_topic
+
+        report = Fixtures.json('consumer_report', symbolize_names: false)
+
+        partition_data = report.dig(*partition_scope)
+
+        partition_data['poll_state'] = 'paused'
+        partition_data['poll_state_ch'] = 1_000_000_000_000
+
+        produce(reports_topic, report.to_json)
+
+        get 'health/changes'
+      end
+
+      it do
+        expect(response).to be_ok
+        expect(body).to include(breadcrumbs)
+        expect(body).not_to include(pagination)
+        expect(body).not_to include(support_message)
+        expect(body).to include('Until manual resume')
+      end
+    end
+  end
 end