+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Karafka Web changelog
 
 ## 0.8.0 (Unreleased)
-- **[Feature]** Track and report pause timeouts.
+- **[Feature]** Track and report pause timeouts via "Changes" view in Health.
 - **[Feature]** Introduce pending jobs visibility alongside of running jobs both in total and per process.
 - **[Feature]** Introduce states migrations for seamless upgrades.
 - **[Feature]** Introduce "Data transfers" chart with data received and data sent to the cluster.
@@ -16,6 +16,8 @@
 - [Enhancement] Support time ranges for graphs in OSS.
 - [Enhancement] Report last poll time for each subscription group.
 - [Enhancement] Show last poll time per consumer instance.
+- [Fix] Fix times precisions that could be incorrectly reported by 1 second in few places.
+- [Fix] Fix random order in Consumers groups Health view.
 - [Change] Rename "Busy" to "Running" to align with "Running Jobs".
 - [Change] Rename "Active subscriptions" to "Subscriptions" as process subscriptions are always active.
 - [Maintenance] Introduce granular subscription group contracts.